32.2.Normalizer & Completely Custom Fields{Fixing the Tests}
To fix the test, change createUserAndLogin() to just createUser()... and remove the first argument. Now use $this->createUserAndLogin() to log in as a totally different user. Now we're making a GET request for the cheeseplease@example.com user data but we're authenticated as this other user. So, we should not see the phoneNumber field

Run the tests again:

php bin/phpunit
And... all green.