32.2.Normalizer & Completely Custom Fields{Adding the Custom isMe Field}
Ok, back to our original mission... which will be delightfully simple: adding a custom isMe field to User. Because  is an array, we can add whatever fields we want. Up here, I'll create a variable called  $isOwner set to what we have in the if statement: $this->userIsOwner(). Now we can use $isOwner in the if and add the custom field: $data['isMe'] = $isOwner.

Et voilà! Test it! Execute the operation again and... there it is: isMe false, false and true! Just remember the downside to this approach: our documentation has no idea that this isMe field exists. If we refresh this page and open the docs for fetching a single User... yep! There's no mention of isMe. Of course, you could add a public function isMe() in User, put it in the user:read group, always return false, then override the isMe key in your normalizer with the real value. That would give you the custom field and the docs. But sheesh... that's... getting kinda hacky.

Next, let's look more at the owner field on CheeseListing. It's interesting: we're currently allowing the user to set this property when they POST to create a CheeseListing. Does that make sense? Or should it be set automatically? And if we do want an API user to be able to send the owner field via the JSON, how do we prevent them from creating a CheeseListing and setting the owner to some other user? It's time to see where security & validation meet.