33.1.Locking down the CheeseListing.owner Field{Making owner only Settable on POST}
Inside CheeseListing, the owner property - that's the related User object that owns this CheeseListing - is required in the database. And, thanks to the cheese:write group, it's a writable field in our API.

In fact, even though I've forgotten to add an @Assert\NotBlank constraint to this property, an API client must send this field when creating a new CheeseListing. They can also change the owner by sending the field via a PUT request. We even added some fanciness where, when you create or edit a CheeseListing, you can modify the owner's username all at the same time. This works because that property is in the cheese:write group... oops. I forgot to change this group when we were refactoring - that's how it should look. The cheese:item:get group means this field will be embedded when we GET a single CheeseListing - that's the "item operation" - and cheese:write means it's writable when using any write operation for CheeseListing. That's some crazy stuff we set up on our previous tutorial.

Removing username Updatable Fanciness

But now, I want to simplify in two ways. First, I only want the owner property to be set when we create the CheeseListing, I don't want it to be changeable. And second, let's get rid of the fanciness of being able to edit the username property via a CheeseListing operation. For that second part, remove the cheese:write group from username. We can now also take off the @Assert\Valid() annotation. This caused the User to be validated during the CheeseListing operations... which was needed to make sure someone didn't set the username to an invalid value while updating a CheeseListing.

Making owner only Settable on POST

Now, how can we make the owner property settable for the POST operation but not the PUT operation?

Open up AutoGroupResourceMetadataFactory. This monster automatically adds three serialization groups in all situations. We can use this last one to include a field only for a specific operation. Change cheese:write to cheese:collection:post. That follows the pattern: "short name", colon, collection, colon, then the operation name: post.

Congratulations, the owner can no longer be changed.