33.2.Locking down the CheeseListing.owner Field{Should Owner Be Set Automatically?}
Should Owner Be Set Automatically?

But... hold up. Isn't it kinda weird that we allow the API client to send the owner field at all? I mean... shouldn't we instead not make owner part of our API and then write some code to automatically set this to the currently-authenticated user?

Um, maybe. Automatically setting the owner property is kinda nice... and it would also make our API easier to use. We will talk about how to do this later. But I don't want to completely remove owner from my API. Why? Well, what if we created an admin interface where admin users could create cheese listings on behalf of other users. In that case, we would want the owner field to be part of our API.

But... hmm if we allow the owner field to be sent... we can't just allow API clients to create a CheeseListing and set the owner to whoever they want. Sure, maybe an admin user should be able to do this... but how can we prevent a normal user from setting the owner to someone else?

Two Ways to Protect a Writable Field

Backing up, if you need to control some behavior around the way a field is set based on the authenticated user, you have two options. First, you could prevent some users from writing to the field entirely. That's done by putting the property into a special serialization group then dynamically adding that group either in a context builder or in a custom normalizer. We've done that already with admin:read and admin:write.

Second, if the field should be writable by all users... but the data that's allowed to be set on the field depends on who is logged in, then the solution is validation.

Here's our goal: prevent an API client from POSTing an owner value that is different than their own IRI... with an exception added for admin users: they can set owner to anyone.

Let's codify this into a test first. Open CheeseListingResourceTest. Inside testCreateCheeseListing(), we're basically verifying that you do need to be logged in to use the operation. We get a 401 before we're authenticated... then after logging in, we get a 400 status code because access will be granted... but our empty data will fail validation.

Let's make this test more interesting! Create a new $authenticatedUser variable set to who we're logged in as. Then create an $otherUser variable set to... another user in the database.

Here's the plan: I want to make another POST request to /api/cheeses with valid data... except that we'll set the owner field to this $otherUser... a user that we are not logged in as. Start by creating a $cheesyData variable set to an array with title, description and price. These are the three required fields other than owner.

Now, copy the request and status code assertion from before, paste down here and set the json to $cheesyData plus the owner property set to /api/users/ and then $otherUser->getId().

In this case, the status code should still be 400 once we've coded all of this: passing the wrong owner will be a validation error. I'll add a little message to the assertion to make it obvious why it's failing:
  not passing the correct owner

I like it! We're logging in as cheeseplease@example.com... then we're trying to create a CheeseListing that's owned by a totally different user. This is the behavior we want to prevent.

While we're here, copy these two lines again and change $otherUser to $authenticatedUser. This should be allowed, so change the assertion to look for the happy 201 status code.

You know the drill: once you've written a test, you get to celebrate by watching it fail! Copy the method name, flip over to your terminal and run:

php bin/phpunit --filter=testCreateCheeseListing
And... it fails!

Failed asserting response status code is 400 - got 201 Created.

So we are currently able to create cheese listings and set the owner as a different user. Cool! Next, let's prevent this with a custom validator.