37.2.Query Extension: Auto-Filter a Collection
Now let's update the test for the behavior that we want. Allow  to stay not published, but publish the other two: $cheeseListing2->setIsPublished(true)... and then paste that below and rename it for .

Once we're done with the feature, we will only want the second two to be returned. Change the assertion to 2. Now we have a failing test.